requirements/base.txt
psycopg2==2.7.3.2
Pillow==2.9.0
Django==1.11.1
six==1.9.0

BeautifulSoup4==4.4.0

django-dotenv==1.3.0
dealer==2.0.5

django-bootstrap3==9.1.0
